LayoutTests/media/modern-media-controls/start-support/start-support-audio.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<audio src="../../content/test.mp4" style="width: 320px; height: 240px;"></audio>
<div id="shadow"></div>
<script type="text/javascript">

description("Testing the <code>StartSupport</code> behavior with an audio element.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("audio");
const mediaController = createControls(shadowRoot, media, null);

shouldBeFalse("mediaController.controls.showsStartButton");

shadowRoot.host.remove();
media.remove();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
