Safari-609.1.11.1/LayoutTests/media/modern-media-controls/start-support/start-support-audio-expected.txt

Testing the StartSupport behavior with an audio element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.showsStartButton is false
PASS successfullyParsed is true

TEST COMPLETE
